// chat_ui/src/models/SequentialPageModel.tsx
/*
 * Copyright 2018. University of Southern California
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *    http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import { ISequentialPageCheckboxModel } from "./SequentialPageCheckboxModel"


export interface ISequentialPageModel {
    readonly id: string
    readonly hashedId: string
    readonly pageId: string
    readonly pageTitle: string
    readonly sentences: string[]
    readonly stepsSentences: string[][]
    readonly stepsImages: string[][]
    stepsSentencesWithImages: string[][]
    jsonToSequentialPageModel?: (jsonObject: object) => ISequentialPageModel
    sequentialPageModelToSequentialPageCheckboxModel?: (section: string, sectionValue: string) => ISequentialPageCheckboxModel

}